clinic/backend/requirements.txt
django==5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
channels==4.0.0
daphne==4.0.0
cryptography==42.0.0
python-dotenv==1.0.0
